requirements.txt
asgiref==3.8.1
cffi==1.16.0
coverage==7.5.4
cryptography==42.0.8
Django==4.2.13
djangorestframework==3.15.2
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
